@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,40 +33,39 @@ dependencies:
     sdk: flutter
 
   # UI
-  cupertino_icons: ^1.0.4
-  fl_chart: ^0.50.6
-  video_player: ^2.4.2
+  cupertino_icons: ^1.0.5
+  fl_chart: ^0.61.0
+  video_player: ^2.6.0
   # Hardware
-  battery_plus: ^2.1.3
-  mobile_scanner: ^1.1.1
+  battery_plus: ^3.0.3
+  mobile_scanner: ^3.0.0
   qr_flutter: ^4.0.0
-  package_info_plus: ^1.4.2
-  camera: ^0.9.7+1
-  qr_code_dart_scan: ^0.4.1
-  connectivity_plus: ^2.3.0
-  device_info_plus: ^3.2.3
-  network_info_plus: ^2.1.3
-  sensors_plus: ^1.3.1
+  package_info_plus: ^3.0.3
+  camera: ^0.10.3+2
+  connectivity_plus: ^3.0.3
+  device_info_plus: ^6.0.0
+  network_info_plus: ^3.0.2
+  sensors_plus: ^2.0.2
   # Utils
-  share_plus: ^4.0.4
-  image_picker: ^0.8.5+3
+  share_plus: ^6.3.1
+  image_picker: ^0.8.7
   # Libraries
   url_strategy: ^0.2.0
-  intl: ^0.17.0
-  screen_brightness: ^0.2.1
-  geolocator: ^8.2.1
-  chewie: ^1.3.3
-  flutter_sms: ^2.3.2
+  intl: ^0.18.0
+  screen_brightness: ^0.2.2
+  geolocator: ^9.0.2
+  chewie: ^1.4.0
+  flutter_sms: ^2.3.3
   sticky_headers: ^0.3.0+2
-  connectivity_plus_web: ^1.2.0
-  package_info_plus_web: ^1.0.5
-  share_plus_web: ^3.0.0
-  device_info_plus_web: ^2.1.0
+  connectivity_plus_web: ^1.2.5
+  package_info_plus_web: ^2.0.0
+  share_plus_web: ^3.1.0
+  device_info_plus_web: ^5.0.0
   network_info_plus_web: ^1.0.1
   battery_plus_web: ^1.1.0
-  sensors_plus_web: ^1.1.0
+  sensors_plus_web: ^1.1.2
   infinite_listview: ^1.1.0
-  grouped_list: ^5.1.1
+  grouped_list: ^5.1.2
 
 dev_dependencies:
   flutter_test:
@@ -77,7 +76,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  lints: ^2.0.0
+  lints: ^2.0.1
   flutter_lints: ^2.0.1
 
 flutter:
