@@ -34,9 +34,9 @@ class _QrCameraPageState2 extends State<QrCameraPage2>
             MobileScanner(
                 fit: BoxFit.contain,
                 // allowDuplicates: false,
-                onDetect: (barcode, args) {
+                onDetect: (barcodes) {
                   setState(() {
-                    _barcode = barcode.rawValue;
+                    _barcode = barcodes.barcodes[0].rawValue;
                   });
                 }),
             Align(
@@ -59,7 +59,7 @@ class _QrCameraPageState2 extends State<QrCameraPage2>
                             overflow: TextOverflow.fade,
                             style: Theme.of(context)
                                 .textTheme
-                                .headline4!
+                                .headlineMedium!
                                 .copyWith(color: Colors.white),
                           ),
                         ),
