@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app_playground/build_config.dart';
 import 'package:flutter_app_playground/ui/home/main_page.dart';
 
 class MyApp extends StatelessWidget {
@@ -8,7 +7,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: !BuildConfig.isRelease,
+      debugShowCheckedModeBanner: false,
       title: 'Flutter Playground',
       theme: ThemeData(
         brightness: Brightness.light,
