@@ -101,7 +101,7 @@ class _GpsPageState extends State<GpsPage> {
     );
 
     return Scaffold(
-      backgroundColor: Theme.of(context).backgroundColor,
+      backgroundColor: Theme.of(context).colorScheme.background,
       body: ListView.builder(
         itemCount: _positionItems.length,
         itemBuilder: (context, index) {
